Avoid consecutive if-statements

+++ daemon98.py
@@ -89,8 +89,7 @@ def do_mv_data(rpath):
     if DEBUG:print "{0} internal locks exist".format(count_internal_locks)
 
   for fname in glob.glob(r'/tmp/' + leaf + '/*.csv'):
-    if os.path.isfile(clientlock):
-      if not (os.path.isfile(rpath + "/" + os.path.split(fname)[1])):
+    if os.path.isfile(clientlock) and not (os.path.isfile(rpath + "/" + os.path.split(fname)[1])):
         if DEBUG:print "moving data " + fname
         shutil.move(fname, rpath)
